rewrite action -> doSetTheme
