+++ ui/js/page/settings/index.js
@@ -5,14 +5,8 @@ import {
   doSetDaemonSetting,
   doSetClientSetting,
   doSetTheme,
-  doGetThemes,
 } from "actions/settings";
-import {
-  selectDaemonSettings,
-  selectShowNsfw,
-  selectThemes,
-  selectTheme,
-} from "selectors/settings";
+import { selectDaemonSettings, selectShowNsfw } from "selectors/settings";
 import SettingsPage from "./view";
 
 const select = state => ({
@@ -25,7 +19,6 @@ const perform = dispatch => ({
   clearCache: () => dispatch(doClearCache()),
   setClientSetting: (key, value) => dispatch(doSetClientSetting(key, value)),
   setTheme: name => dispatch(doSetTheme(name)),
-  getThemes: () => dispatch(doGetThemes()),
 });
 
 export default connect(select, perform)(SettingsPage);
